simple_cli/pyproject.toml
[project]
name = "encryption-at-rest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyring>=25.6.0",
    "pygithub>=2.6.1",
]